fix(worker,cdn): logs order
 * read stdout and stderr in the same goroutine to avoid mixing logs
 * store the number of millisecond for each line of logs as score in redis and retrieve it in "Number" field from a log line in the UI

+++ engine/cdn/cdn_item_test.go
@@ -242,7 +242,7 @@ func TestGetItemValue_ThousandLines(t *testing.T) {
 		UnitID: s.Units.LogsBuffer().ID(),
 	}
 	for i := 0; i < 1000; i++ {
-		require.NoError(t, s.Units.LogsBuffer().Add(iu, uint(i), fmt.Sprintf("Line %d\n", i)))
+		require.NoError(t, s.Units.LogsBuffer().Add(iu, float64(i), fmt.Sprintf("Line %d\n", i)))
 	}
 
 	require.NoError(t, s.completeItem(context.TODO(), db, iu))
@@ -267,9 +267,9 @@ func TestGetItemValue_ThousandLines(t *testing.T) {
 	var lines []redis.Line
 	require.NoError(t, json.Unmarshal(buf.Bytes(), &lines), "given json should be valid")
 	require.Len(t, lines, 200)
-	require.Equal(t, int64(773), lines[0].Number)
+	require.Equal(t, float64(773), lines[0].Number)
 	require.Equal(t, "Line 773\n", lines[0].Value)
-	require.Equal(t, int64(972), lines[199].Number)
+	require.Equal(t, float64(972), lines[199].Number)
 	require.Equal(t, "Line 972\n", lines[199].Value)
 
 	_, _, rc, _, err = s.getItemLogValue(context.Background(), sdk.CDNTypeItemStepLog, apiRefhash, getItemLogOptions{
@@ -284,9 +284,9 @@ func TestGetItemValue_ThousandLines(t *testing.T) {
 
 	require.NoError(t, json.Unmarshal(buf.Bytes(), &lines), "given json should be valid")
 	require.Len(t, lines, 227)
-	require.Equal(t, int64(773), lines[0].Number)
+	require.Equal(t, float64(773), lines[0].Number)
 	require.Equal(t, "Line 773\n", lines[0].Value)
-	require.Equal(t, int64(999), lines[226].Number)
+	require.Equal(t, float64(999), lines[226].Number)
 	require.Equal(t, "Line 999\n", lines[226].Value)
 }
 
@@ -349,7 +349,7 @@ func TestGetItemValue_Reverse(t *testing.T) {
 		UnitID: s.Units.LogsBuffer().ID(),
 	}
 	for i := 0; i < 5; i++ {
-		require.NoError(t, s.Units.LogsBuffer().Add(iu, uint(i), fmt.Sprintf("Line %d\n", i)))
+		require.NoError(t, s.Units.LogsBuffer().Add(iu, float64(i), fmt.Sprintf("Line %d\n", i)))
 	}
 
 	require.NoError(t, s.completeItem(context.TODO(), db, iu))
@@ -373,9 +373,9 @@ func TestGetItemValue_Reverse(t *testing.T) {
 	var lines []redis.Line
 	require.NoError(t, json.Unmarshal(buf.Bytes(), &lines), "given json should be valid")
 	require.Len(t, lines, 5)
-	require.Equal(t, int64(4), lines[0].Number)
+	require.Equal(t, float64(4), lines[0].Number)
 	require.Equal(t, "Line 4\n", lines[0].Value)
-	require.Equal(t, int64(0), lines[4].Number)
+	require.Equal(t, float64(0), lines[4].Number)
 	require.Equal(t, "Line 0\n", lines[4].Value)
 
 	// Get From Buffer
@@ -393,9 +393,9 @@ func TestGetItemValue_Reverse(t *testing.T) {
 
 	require.NoError(t, json.Unmarshal(buf.Bytes(), &lines), "given json should be valid")
 	require.Len(t, lines, 2)
-	require.Equal(t, int64(2), lines[0].Number)
+	require.Equal(t, float64(2), lines[0].Number)
 	require.Equal(t, "Line 2\n", lines[0].Value)
-	require.Equal(t, int64(1), lines[1].Number)
+	require.Equal(t, float64(1), lines[1].Number)
 	require.Equal(t, "Line 1\n", lines[1].Value)
 }
 
@@ -459,7 +459,7 @@ func TestGetItemValue_ThousandLinesReverse(t *testing.T) {
 		UnitID: s.Units.LogsBuffer().ID(),
 	}
 	for i := 0; i < 1000; i++ {
-		require.NoError(t, s.Units.LogsBuffer().Add(iu, uint(i), fmt.Sprintf("Line %d\n", i)))
+		require.NoError(t, s.Units.LogsBuffer().Add(iu, float64(i), fmt.Sprintf("Line %d\n", i)))
 	}
 
 	require.NoError(t, s.completeItem(context.TODO(), db, iu))
@@ -486,9 +486,9 @@ func TestGetItemValue_ThousandLinesReverse(t *testing.T) {
 	var lines []redis.Line
 	require.NoError(t, json.Unmarshal(buf.Bytes(), &lines), "given json should be valid")
 	require.Len(t, lines, 200)
-	require.Equal(t, int64(226), lines[0].Number)
+	require.Equal(t, float64(226), lines[0].Number)
 	require.Equal(t, "Line 226\n", lines[0].Value)
-	require.Equal(t, int64(27), lines[199].Number)
+	require.Equal(t, float64(27), lines[199].Number)
 	require.Equal(t, "Line 27\n", lines[199].Value)
 
 	_, _, rc, _, err = s.getItemLogValue(context.Background(), sdk.CDNTypeItemStepLog, apiRefhash, getItemLogOptions{
@@ -504,8 +504,8 @@ func TestGetItemValue_ThousandLinesReverse(t *testing.T) {
 
 	require.NoError(t, json.Unmarshal(buf.Bytes(), &lines), "given json should be valid")
 	require.Equal(t, len(lines), 227)
-	require.Equal(t, int64(226), lines[0].Number)
+	require.Equal(t, float64(226), lines[0].Number)
 	require.Equal(t, "Line 226\n", lines[0].Value)
-	require.Equal(t, int64(0), lines[226].Number)
+	require.Equal(t, float64(0), lines[226].Number)
 	require.Equal(t, "Line 0\n", lines[226].Value)
 }
